Adding makefile, moving .h to include folder

+++ lymui/helper.c
@@ -6,8 +6,9 @@
 //  Copyright © 2018 Marc. All rights reserved.
 //
 
-#include "helper.h"
 #include <stdlib.h>
+#include <stdint.h>
+#include "helper.h"
 
 // Global variable of a list of RgbHex
 struct RgbHex *list;
@@ -19,7 +20,7 @@ char * ushortToHex(uint8_t v) {
     
     char HEX[] = "0123456789ABCDEF";
     char *hexa = malloc(1);
-    ushort accum = 2;
+    uint8_t accum = 2;
     while(accum > 0) {
         hexa[accum - 1] = HEX[(v % 16)];
         v = v / 16;
